requirements.txt
pandas
numpy
# psycopg2
psycopg2-binary
datetime
SQLAlchemy
pydantic
python-dotenv
transformers
torch
mlflow
boto3==1.34.103
psutil
tqdm